requirements.txt
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-ratelimit==3.0.0
